Fix missing open line between multi-line let-binding with poly-typexpr
The is_simple check for let bindings only considers the location of
the body exp. This location happens to usually be large enough. But
for one of a group of let bindings which has an explicit polymorphic
type constraint, the location of the body exp contains only the body,
and does not include the location of the pattern. This means that
cases such as:
```ocaml
[@@@ocamlformat "let-binding-spacing=compact"]

let f x = x

and g : 'a. (_ -> _ -> _ -> 'a) -> _ -> _ -> _ -> 'a =
 fun h a b -> h (i a) (i b) (i c)
```
get misformatted since `g` is misidentified as being a one-liner, and
 the open line between `f` and `g` gets removed.

This patch fixes this by using the location of the entire let-binding
instead of the body expression.